Rename `EventsPage` to `UpcomingEvents`
`UpcomingEvents` more accurately reflects the page itself and removes
the redudant `Page` suffix.
